These line of code i used for clean the data .
This Jupyter Notebook is the starting point for an exploratory data analysis (EDA) project on Netflix content. It leverages Python and the Pandas library  to load and begin examining the titles.csv dataset, and i used here ast to makes the column cleaner and easier to work with for analysis and visualization. which includes metadata about Netflix's movie and TV show catalog.